update `cc` crate
(and a few lints added in the last 9 months)

most importantly this gets us
https://github.com/rust-lang/cc-rs/pull/1496 which fixes
https://github.com/oxidecomputer/helios/issues/192 -style
miscompilation of `crates/tofino/src/c/pci.c`.

I don't this is likely to be an issue in practice since it seems like
this code would only be used at startup of programs depending on
`tofino`, if at all. But getting the `errno` symbol out of `dpd` and
`nexus` is a proof that helios#192 won't be a nuisance.